Refactor helpers.php, PayController.php, validation.js, Response.php, styles.css, and DiscountController.php files

+++ app/services/PromotionService.php
@@ -3,6 +3,9 @@
 namespace App\Services;
 
 use App\Core\Database\Database;
+use App\Core\Database\QueryBuilder;
+use App\Core\Logger;
+use App\Core\Request;
 use App\Dtos\JsonDataErrorRespose;
 use App\Dtos\JsonResponse;
 use App\Dtos\TrangThaiKhuyenMai;
@@ -12,9 +15,24 @@ class PromotionService
 
     public static function getPromotionByCode($code)
     {
-        $query = "SELECT * FROM KhuyenMai WHERE MaKhuyenMai = ?";
+        $query = "SELECT * FROM KhuyenMai WHERE BINARY MaKhuyenMai = ?";
         $promotion = Database::queryOne($query, [$code]);
         return $promotion;
+    }
+    public static function toggleHidePromotion($id)
+    {
+        $promotion = self::getPromotionByCode($id);
+        if (!$promotion) {
+            return JsonResponse::error("Không tìm thấy mã khuyến mãi");
+        }
+        $newStatus = $promotion['TrangThai'] == TrangThaiKhuyenMai::An->value ? TrangThaiKhuyenMai::DangHoatDong->value : TrangThaiKhuyenMai::An->value;
+        $result = Database::update('KhuyenMai', ['TrangThai' => $newStatus], "MaKhuyenMai='$id'");
+        if ($result) {
+            return JsonResponse::ok();
+        } else {
+            return JsonResponse::error('Cập nhật thất bại', 500);
+        }
+
     }
     public static function getAllDiscount()
     {
@@ -28,20 +46,29 @@ class PromotionService
         $discount = Database::query($sql, []);
         return $discount;
     }
+
     public static function createNewDiscount($data)
     {
         $params = [
             'MaKhuyenMai' => $data['MaKhuyenMai'],
             'TenKhuyenMai' => $data['TenKhuyenMai'],
             'MoTa' => $data['MoTa'],
-            'NgayBatDau' => $data['NgayBatDau'],
-            'NgayKetThuc' => $data['NgayKetThuc'],
-            'GiamToiDa' => $data['GiamToiDa'],
-            'TiLeGiam' => $data['TiLeGiam']
+            'NgayBatDau' => getArrayValueSafe($data, 'NgayBatDau', null),
+            'NgayKetThuc' => getArrayValueSafe($data, 'NgayKetThuc', null),
+            'GiamToiDa' => getArrayValueSafe($data, 'GiamToiDa', null),
+            'GiaTriGiam' => $data['GiaTriGiam'],
+            'GioiHanSuDung' => getArrayValueSafe($data, 'GioiHanSuDung', null),
+            'GioiHanTrenKhachHang' => getArrayValueSafe($data, 'GioiHanTrenKhachHang', null),
+            'GiaTriToiThieu' => getArrayValueSafe($data, 'GiaTriToiThieu', null),
+            'TrangThai' => getArrayValueSafe($data, 'TrangThai', TrangThaiKhuyenMai::DangHoatDong->value),
+            'MaLoaiVe' => getArrayValueSafe($data, 'MaLoaiVe', null),
+            'DiemToiThieu' => getArrayValueSafe($data, 'DiemToiThieu', 0)
         ];
         $result = Database::insert('KhuyenMai', $params);
         if ($result) {
-            return JsonResponse::ok();
+            return JsonResponse::ok([
+                'MaKhuyenMai' => $data['MaKhuyenMai'],
+            ]);
         }
         return JsonResponse::error('Thêm mới thất bại', 500);
     }
@@ -51,12 +78,18 @@ class PromotionService
         $params = [
             'TenKhuyenMai' => $data['TenKhuyenMai'],
             'MoTa' => $data['MoTa'],
-            'NgayBatDau' => $data['NgayBatDau'],
-            'NgayKetThuc' => $data['NgayKetThuc'],
-            'GiamToiDa' => $data['GiamToiDa'],
-            'TiLeGiam' => $data['TiLeGiam']
+            'NgayBatDau' => getArrayValueSafe($data, 'NgayBatDau', null),
+            'NgayKetThuc' => getArrayValueSafe($data, 'NgayKetThuc', null),
+            'GiamToiDa' => getArrayValueSafe($data, 'GiamToiDa', null),
+            'GiaTriGiam' => $data['GiaTriGiam'],
+            'GioiHanSuDung' => getArrayValueSafe($data, 'GioiHanSuDung', null),
+            'GioiHanTrenKhachHang' => getArrayValueSafe($data, 'GioiHanTrenKhachHang', null),
+            'GiaTriToiThieu' => getArrayValueSafe($data, 'GiaTriToiThieu', null),
+            'TrangThai' => getArrayValueSafe($data, 'TrangThai', TrangThaiKhuyenMai::DangHoatDong->value),
+            'MaLoaiVe' => getArrayValueSafe($data, 'MaLoaiVe', null),
+            'DiemToiThieu' => getArrayValueSafe($data, 'DiemToiThieu', 0)
         ];
-        $result = Database::update('KhuyenMai', $params, "MaKhuyenMai=$id");
+        $result = Database::update('KhuyenMai', $params, "MaKhuyenMai='$id'");
         if ($result) {
             return JsonResponse::ok();
         }
@@ -65,6 +98,12 @@ class PromotionService
 
     public static function deleteDiscount($id)
     {
+        // check if any bill use this discount
+        $query = "SELECT COUNT(*) as count FROM HoaDon WHERE MaKhuyenMai = ?";
+        $count = Database::queryOne($query, [$id]);
+        if ($count['count'] > 0) {
+            return JsonResponse::error("Không thể xóa khuyến mãi này vì đã có hóa đơn sử dụng", 409);
+        }
         $result = Database::delete('KhuyenMai', "MaKhuyenMai=$id");
         if ($result) {
             return JsonResponse::ok();
@@ -79,7 +118,7 @@ class PromotionService
         if (!$promotion) {
             return JsonResponse::error("Không tìm thấy mã khuyến mãi");
         }
-        if ($promotion['TrangThai'] != null && $promotion['TrangThai'] == TrangThaiKhuyenMai::An) {
+        if ($promotion['TrangThai'] != null && $promotion['TrangThai'] == TrangThaiKhuyenMai::An->value) {
             return JsonResponse::error("Khuyến mãi không còn hoạt động");
         }
         if ($promotion['NgayBatDau'] != null && $promotion['NgayBatDau'] > date('Y-m-d H:i:s')) {
@@ -104,14 +143,15 @@ class PromotionService
             return JsonResponse::error("Tổng giá trị vé phải lớn hơn " . $promotion["GiaTriToiThieu"]);
         }
         $reducePrice = 0;
-        $discountValue = $promotion['GiaTriGiam'];
+        $discountValue = intval($promotion['GiaTriGiam']);
         if ($discountValue > 100) {
             $reducePrice = $discountValue;
         } else {
+
             $reducePrice = $totalPrice * $discountValue / 100;
             $maxDiscount = $promotion['GiamToiDa'];
             if ($maxDiscount != null && $maxDiscount > 0) {
-                $reducePrice = min($discountValue, $maxDiscount);
+                $reducePrice = min($reducePrice, $maxDiscount);
             }
         }
         $finalReducePrice = min($reducePrice, $totalPrice);
@@ -141,5 +181,44 @@ class PromotionService
         ], " MaKhuyenMai = '" . $promotion['MaKhuyenMai'] . "'");
         return JsonResponse::ok();
 
+    }
+
+    public static function getAllPromotions($query = [])
+    {
+        $queryBuilder = new QueryBuilder();
+        $queryBuilder
+            ->select(['*'])->from('KhuyenMai')
+            ->where("1", '=', '1');
+        $keyword = getArrayValueSafe($query, 'tu-khoa', '');
+        $khoangThoiGianTu = getArrayValueSafe($query, 'khoang-thoi-gian-tu', '');
+        $khoangThoiGianDen = getArrayValueSafe($query, 'khoang-thoi-gian-den', '');
+        $trangThais = getArrayValueSafe($query, 'trang-thais', []);
+        $page = getArrayValueSafe($query, 'trang', 1);
+        $limit = getArrayValueSafe($query, 'limit', 20);
+        $offset = ($page - 1) * $limit;
+        $orderBy = getArrayValueSafe($query, 'sap-xep', 'MaKhuyenMai');
+        $orderType = getArrayValueSafe($query, 'thu-tu', 'ASC');
+
+
+        if ($keyword != '') {
+            $queryBuilder->and();
+            $queryBuilder->startGroup();
+            $queryBuilder->where('MaKhuyenMai', 'LIKE', "%$keyword%");
+            $queryBuilder->orWhere('TenKhuyenMai', 'LIKE', "%$keyword%");
+            $queryBuilder->endGroup();
+        }
+        if ($khoangThoiGianTu != '') {
+            $queryBuilder->andWhere('NgayBatDau', '>=', $khoangThoiGianTu);
+        }
+        if ($khoangThoiGianDen != '') {
+            $queryBuilder->andWhere('NgayKetThuc', '<=', $khoangThoiGianDen);
+        }
+        if ($trangThais) {
+            $queryBuilder->andWhere('TrangThai', 'IN', $trangThais);
+        }
+        Request::setQueryCount($queryBuilder->count());
+        $queryBuilder->orderBy($orderBy, $orderType);
+        $queryBuilder->limit($limit, $offset);
+        return $queryBuilder->get();
     }
 }
